Consolidated a few files and started building up the parser.
Token types are now organized under the LexicalSpec.cs file.
Defined the basic types that make up a Grammar.
Renamed the XML tags in the Grammar.xml file.
Stubbed out an IParser interface and DefaultParser class.
