--- ClarionLinter/Language/ILexer.cs
@@ -1,30 +1,0 @@
-﻿namespace Language
-{
-
-    /// <summary>
-    /// ILexer represents a stream of Lexemes.
-    /// </summary>
-    public interface ILexer
-    {
-
-        /// <summary>
-        /// Checks whether another Lexeme exists in the stream.
-        /// </summary>
-        /// <returns>Whether another Lexeme exists.</returns>
-        bool HasNext();
-
-        /// <summary>
-        /// Retrieves the next Lexeme without removing it from the stream.
-        /// </summary>
-        /// <returns>The next Lexeme in the stream.</returns>
-        Lexeme Peek();
-
-        /// <summary>
-        /// Retrieves and removes the next Lexeme in the stream.
-        /// </summary>
-        /// <returns>The next Lexeme in the stream.</returns>
-        Lexeme Read();
-
-    }
-
-}
