+++ ClarionLinter/Language/Parser.cs
@@ -1,0 +1,25 @@
+﻿namespace ClarionLinter.Language
+{
+
+    /// <summary>
+    /// IParser represents an abstract syntax tree.
+    /// </summary>
+    public interface IParser
+    {
+
+        // TODO: I want to be able implement IParser to create decorators that will be used to
+        // 
+
+    }
+
+    /// <summary>
+    /// A DefaultParser parses a programming language from a stream of Lexemes.
+    /// </summary>
+    public class DefaultParser : IParser
+    {
+
+        // TODO: Implement IParser with a multi-pass bottom-up parsing algorithm.
+
+    }
+
+}
